Core: Fix gcc warning about signed/unsigned integer comparison
Fix warning: 'comparison between signed and unsigned integer expressions'

+++ src/corelib/kernel/qtranslator.cpp
@@ -782,7 +782,7 @@ bool QTranslatorPrivate::do_load(const uchar *data, int len, const QString &dire
         data += 4;
         if (!tag || !blockLen)
             break;
-        if (end - data < blockLen) {
+        if (quint32(end - data) < blockLen) {
             ok = false;
             break;
         }
